refactor: Consolidate authentication schemas into authSchema and update references
